% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adagrad.R
\name{adagrad}
\alias{adagrad}
\title{Optimizarea stochastica AdaGrad}
\usage{
adagrad(X, y, w, eta, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul de ponderi}

\item{eta}{este rata de invatare}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea stochastica AdaGrad
}
