% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\name{adam}
\alias{adam}
\title{Optimizarea l1-regularizata Lasso}
\usage{
adam(X, y, w, alpha, beta1, beta2, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul initial de ponderi}

\item{alpha}{este rata de invatare}

\item{beta1}{este rata de descompunere exponentiala pentru primul moment estimat (valoarea recomandata este 0.9)}

\item{beta2}{este rata de descompunere exponentiala pentru al doilea moment si este indicat sa fie apropiata de 1.0}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea l1-regularizata Lasso
}
