% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_linear_regression}
\alias{fit_linear_regression}
\title{Functia de antrenare a unui model liniar}
\usage{
fit_linear_regression(
  X,
  y,
  w_init = rep(0, dim(X)[2]),
  optimizer = "gd",
  learning_rate = 0.001,
  model_train_error = 1e-05,
  newton_lambda = 10,
  adam_alpha = 0.01,
  adam_beta1 = 0.9,
  adam_beta2 = 0.999,
  verbose = FALSE
)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul}

\item{w_init}{este vectorul initial de ponderi (valoare default)}

\item{optimizer}{este algoritmul de optimizare utilizat}

\item{learning_rate}{este rata de invatare pentru algoritmul de optimizare}

\item{model_train_error}{este performanta pe care o asteptam de la model}

\item{newton_lambda}{este constanta care asigura regularizare in metoda Newton}

\item{adam_alpha}{este parametrul alpha pentru algoritmul ADAM}

\item{adam_beta1}{este parametrul beta1 pentru algoritmul ADAM}

\item{adam_beta2}{este parametrul beta2 pentru algoritmul ADAM}

\item{verbose}{este parametrul care, setat la TRUE, afiseaza detalii despre antrenare}
}
\description{
Functia de antrenare a unui model liniar
}
\examples{
fit_linear_regression(X, y, optimizer = "adam", verbose = TRUE)
}
