% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_ridge_regression}
\alias{fit_ridge_regression}
\title{Functia de antrenare a unui model Ridge}
\usage{
fit_ridge_regression(
  X,
  y,
  w = rep(0, dim(X)[2]),
  learning_rate = 0.001,
  lambda = 10,
  model_train_error = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul}

\item{learning_rate}{este rata de invatare pentru algoritmul de optimizare}

\item{lambda}{este constanta de regularizare pentru regularizarea l2}

\item{model_train_error}{este performanta pe care o asteptam de la model}

\item{verbose}{este parametrul care, setat la TRUE, afiseaza detalii despre antrenare}

\item{w_init}{este vectorul initial de ponderi (valoare default)}
}
\description{
Functia de antrenare a unui model Ridge
}
\examples{
fit_ridge_regression(X, y, verbose = TRUE)
}
