% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Optimizarea l1-regularizata Lasso}
\usage{
lasso(X, y, w, eta, lambda, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul initial de ponderi}

\item{eta}{este rata de invatare}

\item{lambda}{este constanta utilizata in algoritm pentru intervalul centrat de relevanta}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea l1-regularizata Lasso
}
