% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton.R
\name{newton}
\alias{newton}
\title{Optimizarea Newton cu Hessiana fixata}
\usage{
newton(X, y, w, lambda, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul de ponderi}

\item{lambda}{este constanta care asigura inversabilitatea hessianei, este strict pozitiva}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea Newton cu Hessiana fixata
}
