% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Optimizarea l2-regularizata Ridge}
\usage{
ridge(X, y, w, eta, lambda, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul initial de ponderi}

\item{eta}{este rata de invatare}

\item{lambda}{este constanta de regularizare si este strict pozitiva}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea l2-regularizata Ridge
}
