% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_gradient_descent.R
\name{stochastic_gradient_descent}
\alias{stochastic_gradient_descent}
\title{Optimizarea stochastica pentru Gradient Descent}
\usage{
stochastic_gradient_descent(X, y, w, eta, eps, verbose)
}
\arguments{
\item{X}{este setul de date}

\item{y}{este labelul asociat}

\item{w}{este vectorul initial de ponderi}

\item{eta}{este rata de invatare}

\item{eps}{este performanta asteptata a antrenarii}

\item{verbose}{este parametrul care, setat la TRUE, ofera detalii in timp real despre implementare, dar numai pentru vectori din X bidimensionali}
}
\description{
Optimizarea stochastica pentru Gradient Descent
}
